+++ domain/src/main/java/com/agileburo/anytype/domain/page/navigation/PageLink.kt
@@ -6,14 +6,12 @@ import com.agileburo.anytype.domain.block.model.Block
  * @property id page id
  * @property fields page fields
  * @property snippet text from first child block of the page
- * @property lastOpened last time opened
  * @property hasInboundLinks does this page has inbound pages
  */
 data class PageInfo(
     val id: String,
     val fields: Block.Fields,
     val snippet: String?,
-    val lastOpened: Long,
     val hasInboundLinks: Boolean
 )